Working on new TransportTuple.hash implementation
This branch should be applied on `webrtcserver` branch or in `v3` once the former is merged.

Not happy at all yet because:

- `SetHash()` implementation in this branch is too simple. We should consider local IP:port and protocol as well, but we cannot just "sum" those numbers since there is a chance to collide.
- And if we collide (if we have 2 tuples with same `hash` in the same `WebRtcServer`) then we'll have an assertion problem:
  ```
  RTC::WebRtcServer::OnWebRtcTransportTransportTupleRemoved() | failed assertion `this->mapTupleWebRtcTransport.find(tuple->hash) != this->mapTupleWebRtcTransport.end()': tuple not handled
  ```
- So IMHO we need `hash` to be  `int64_t` and we should concatenate remote IP:port first then local IP:port plus protocol somewhere. But summing protocol cannot be just about doing + 1 or +2. imagine this:
  ```
  - Client A connects from udp:1.2.3.4:5001 to udp:9.9.9.9:9000.
  - Client B connects from tcp:1.2.3.4:5000 to tcp:9.9.9.9:9000.
  - If tcp protocol just means "adding 1 to the hash" then both hashes will be the same (BUMPP!).

+++ worker/include/RTC/WebRtcServer.hpp
@@ -100,7 +100,7 @@ namespace RTC
 		// Map of WebRtcTransports indexed by local ICE usernameFragment.
 		absl::flat_hash_map<std::string, RTC::WebRtcTransport*> mapLocalIceUsernameFragmentWebRtcTransport;
 		// Map of WebRtcTransports indexed by TransportTuple.hash.
-		absl::flat_hash_map<size_t, RTC::WebRtcTransport*> mapTupleWebRtcTransport;
+		absl::flat_hash_map<uint32_t, RTC::WebRtcTransport*> mapTupleWebRtcTransport;
 	};
 } // namespace RTC
